+++ packages/env/lib/cli.js
@@ -131,6 +131,14 @@ module.exports = function cli() {
 		},
 		withSpinner( env.run )
 	);
+	yargs.command(
+		'destroy',
+		wpRed(
+			'Destroy the WordPress environment. Delete docker containers and remove local files.'
+		),
+		() => {},
+		withSpinner( env.destroy )
+	);
 	yargs.example(
 		'$0 run cli wp user list',
 		'Runs `wp user list` wp-cli command which lists WordPress users.'