+++ packages/env/lib/commands/index.js
@@ -5,10 +5,12 @@ const start = require( './start' );
 const stop = require( './stop' );
 const clean = require( './clean' );
 const run = require( './run' );
+const destroy = require( './destroy' );
 
 module.exports = {
 	start,
 	stop,
 	clean,
 	run,
+	destroy,
 };